Week 2 part 3

+++ students/steven/Canchu/models/friendModel.js
@@ -33,26 +33,29 @@ module.exports = {
             const sql_friend = 'INSERT INTO friendship (user_id, status, friend_id) VALUES (?,?,?)'
             db.query(sql_friend, [my_id,'pending',friend_id], (error, results) => {
                 if (error) {
-                    console.error('Database error:', error);
-                    return res.status(500).json({ error: 'Server error' });
+                    if (error.code === 'ER_DUP_ENTRY') {
+                        console.error('Duplicate Request Error', error.code);
+                        return res.status(403).json({ error: 'You have already send a request to this person.' });
+                    } else {
+                        console.error('Database error:',error.code)
+                        return res.status(500).json({ error: 'Server error' });
+                    }
                 } else {
-
                     const resultId = results.insertId
-		    const sql_event = 'INSERT INTO event (user_id,type,summary) VALUES (?,?,?)'
+		            const sql_event = 'INSERT INTO event (user_id,type,summary) VALUES (?,?,?)'
                     db.query(sql_event, [my_id,'friend_request',`ID ${friend_id} invited you to be friends.`], (error, results) => {
                         if (error) {
                             console.error('Database error:', error);
                             return res.status(500).json({ error: 'Server error' });
                         } else {
-                            
-                        }
-                        res.status(200).json({
-                            data: {
-                                friendship: {
-                                    id: resultId
+                            res.status(200).json({
+                                data: {
+                                    friendship: {
+                                        id: resultId
+                                    }
                                 }
-                            }
-                        });
+                            })
+                        }
                     })
                 }
             })
@@ -65,11 +68,13 @@ module.exports = {
                 console.error('Database error:', error);
                 return res.status(500).json({ error: 'Server error' });
             }
+            if(results[0] === undefined){
+                return res.status(403).json({ error: 'User ID not existed' });
+            }
             if(results[0].user_id !== self_id && results[0].friend_id !== self_id){
                 console.error("要求刪除Request的ID為",self_id,"，但能同意的只有ID為",results[0].user_id,"或ID為",results[0].friend_id,"的使用者，id為",id)
-		        return res.status(400).json({ error: 'This user has no permission to agree friend request'})
+		        return res.status(400).json({ error: 'This user has no permission to delete friend request'})
             } else {
-                console.log("通過。接下來...")
                 const sql = 'DELETE FROM friendship WHERE id = ?'
                 db.query(sql, [id], (error, results) => {
                     if (error) {
@@ -93,8 +98,11 @@ module.exports = {
             if (error) {
                 console.error('Database error:', error);
                 return res.status(500).json({ error: 'Server error' });
-	    }
-	    console.log("印出來：",id,results,results[0])
+	        }
+	        console.log("印出來：",id,results,results[0])
+            if(results[0] === undefined){
+                return res.status(403).json({ error: 'User ID not existed' });
+            }
             if(results[0].user_id !== self_id){
                 console.error("要求同意Request的ID為",self_id,"，但能同意的只有ID為",results[0].user_id,"的使用者，id為",id)
 		        return res.status(400).json({ error: 'This user has no permission to agree friend request'})
@@ -107,8 +115,9 @@ module.exports = {
                         console.error('Database error:', error);
                         return res.status(500).json({ error: 'Server error' });
                     } else {
-                        const sql = 'INSERT INTO event (user_id,type,summary) VALUES (?,?,?)'
-                        db.query(sql, [friend_id,'friend_request',`ID ${friend_id} has accepted your friend request.`], (error, results) => {
+                        const sql = 'INSERT INTO event (user_id,type,summary,created_at) VALUES (?,?,?,?)'
+                        const eventTime = new Date().toISOString().slice(0, 19).replace('T', ' ');
+                        db.query(sql, [friend_id,'friend_request',`ID ${friend_id} has accepted your friend request.`,eventTime], (error, results) => {
                             if (error) {
                                 console.error('Database error:', error);
                                 return res.status(500).json({ error: 'Server error' });
